refactoring prometheus.go@
