+++ internal/helper.go
@@ -56,62 +56,3 @@ func GetHeadBucket(client *s3.S3, bucketName string) error {
 	log.Printf("Bucket %s exists and is accessible\n", bucketName)
 	return nil
 }
-
-// func addTagsToBucket(client *s3.S3, bucketName string) {
-// 	tags := []*s3.Tag{}
-
-// 	switch bucketName {
-// 	case "nbc-bucket01-logs":
-// 		tags = []*s3.Tag{
-// 			{
-// 				Key:   aws.String("Tenant"),
-// 				Value: aws.String("Niedersachsen"),
-// 			},
-// 		}
-// 	case "dbp-test-bucketlogs":
-// 		tags = []*s3.Tag{
-// 			{
-// 				Key:   aws.String("Tenant"),
-// 				Value: aws.String("Brandenburg"),
-// 			},
-// 		}
-// 	case "dbp-test4logbucket":
-// 		tags = []*s3.Tag{
-// 			{
-// 				Key:   aws.String("Tenant"),
-// 				Value: aws.String("Thueringen"),
-// 			},
-// 		}
-// 	case "dbp-test5-logbucket":
-// 		tags = []*s3.Tag{
-// 			{
-// 				Key:   aws.String("Tenant"),
-// 				Value: aws.String("HPIBosscloud"),
-// 			},
-// 		}
-// 	default:
-// 		tags = []*s3.Tag{
-
-// 			{
-// 				Key:   aws.String("Enviroment"),
-// 				Value: aws.String("Production"),
-// 			},
-// 			{
-// 				Key:   aws.String("Namespace"),
-// 				Value: aws.String("Some Namespace"),
-// 			},
-// 		}
-// 	}
-// 	input := &s3.PutBucketTaggingInput{
-// 		Bucket: aws.String(bucketName),
-// 		Tagging: &s3.Tagging{
-// 			TagSet: tags,
-// 		},
-// 	}
-// 	_, err := client.PutBucketTagging(input)
-// 	if err != nil {
-// 		log.Printf("Error adding tags to bucket %s: %v\n", bucketName, err)
-// 	} else {
-// 		fmt.Printf("Successfully added tags to bucekt %s\n", bucketName)
-// 	}
-// }
